Month Translation, not paid highlighting Fixed the Translation of Bill-month, Not paid bills will be highlighted in index, #26

+++ db/migrate/20130929153443_create_bills.rb
@@ -5,7 +5,7 @@ class CreateBills < ActiveRecord::Migration
       t.date :date, :null => false
       t.decimal :month
       t.decimal :year
-      t.integer :amount, :precision => 10, :scale => 2, :null => false
+      t.decimal :amount, :precision => 10, :scale => 2, :null => false
       t.boolean :paid
       t.text :comment
       t.integer :customer_id, :null => false
